feat!: Rule mechanism for seperated frontend/backend
BREAKING CHANGE: About everything
