+++ src/lib.rs
@@ -1,10 +1,5 @@
-mod cache;
-
-pub mod backend;
-mod cached_function;
-pub mod global_errors;
-pub mod prelude;
-pub mod rules;
 pub mod simulation;
-pub mod state;
-pub mod units;
+mod cached_function;
+pub mod prelude;
+pub mod models;
+mod hash;
