+++ src/simulation.rs
@@ -1,423 +1,587 @@
 use std::{
-    fmt::{Debug, Display},
+    fmt::Debug,
     hash::Hash,
+    sync::{Arc, Mutex},
 };
 
 use crate::prelude::*;
+use hashbrown::HashMap;
+use petgraph::{graph::Graph, visit::EdgeRef};
+use rayon::prelude::*;
 
-use hashbrown::{HashMap, HashSet};
-use petgraph::Graph;
-use serde::{Deserialize, Serialize};
+type StateHash = u64;
+type KnownStates<S> = HashMap<StateHash, S>;
 
-#[derive(Clone, Default)]
-pub struct Step<T> {
-    reachable_states: ReachableStates,
-    applied_rules: HashMap<RuleName, Rule<T>>,
+type TransitionHash = u64;
+type KnownTransitions<T> = HashMap<TransitionHash, T>;
+
+type StateTransitionGraph = Graph<StateHash, (TransitionHash, Probability)>;
+
+pub type StateTransitionGenerator<S, T> =
+    Arc<dyn Fn(S) -> OutgoingTransitions<S, T> + Send + Sync + 'static>;
+
+pub type StateProbabilityDistribution<S> = HashMap<S, Probability>;
+
+pub type OutgoingTransitions<S, T> = Vec<(S, T, Probability)>;
+
+type HashedStateProbabilityDistribution = HashMap<StateHash, Probability>;
+
+pub type Probability = f64;
+pub type Time = u64;
+
+#[derive(Clone)]
+pub struct Simulation<S, T> {
+    state_transition_graph: StateTransitionGraph,
+    probability_distributions: HashMap<Time, HashedStateProbabilityDistribution>,
+    known_states: KnownStates<S>,
+    known_transitions: KnownTransitions<T>,
+    state_transition_generator: CachedFunction<S, OutgoingTransitions<S, T>>,
 }
 
-impl<T> Step<T> {
-    pub fn new(
-        reachable_states: ReachableStates,
-        applied_rules: HashMap<RuleName, Rule<T>>,
-    ) -> Step<T> {
-        Step {
-            reachable_states,
-            applied_rules,
-        }
-    }
-    pub fn reachable_states(&self) -> &ReachableStates {
-        &self.reachable_states
-    }
-    pub fn applied_rules(&self) -> &HashMap<RuleName, Rule<T>> {
-        &self.applied_rules
-    }
-}
-
-#[derive(Clone, Default)]
-pub struct History<T> {
-    steps: Vec<Step<T>>,
-}
-
-impl<T> History<T> {
-    pub fn new(reachable_states: ReachableStates) -> History<T> {
-        History {
-            steps: vec![Step::new(reachable_states, HashMap::new())],
-        }
-    }
-    pub fn steps(&self) -> &Vec<Step<T>> {
-        &self.steps
-    }
-    pub fn time(&self, time: usize) -> Option<&Step<T>> {
-        self.steps.get(time)
-    }
-    pub fn append(&mut self, step: Step<T>) {
-        self.steps.push(step);
-    }
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-struct SerializableStep {
-    reachable_states: ReachableStates,
-    applied_rules: HashSet<RuleName>,
-}
-
-impl SerializableStep {
-    pub fn to_step<T: Clone>(
-        &self,
-        rules: &HashMap<RuleName, Rule<T>>,
-    ) -> Result<Step<T>, RuleError>
-    where
-        T: Hash
-            + Clone
-            + PartialEq
-            + Debug
-            + Default
-            + Serialize
-            + Send
-            + Sync
-            + for<'a> Deserialize<'a>,
-    {
-        Ok(Step {
-            reachable_states: self.reachable_states.clone(),
-            applied_rules: self
-                .applied_rules
-                .iter()
-                .map(|rule_name| {
-                    let rule = rules.get(rule_name);
-                    if let Some(rule) = rule {
-                        Ok((rule_name.clone(), rule.clone()))
-                    } else {
-                        Err(RuleError::RuleNotFound {
-                            rule_name: rule_name.clone(),
-                            context: get_backtrace(),
-                        })
-                    }
-                })
-                .collect::<Result<HashMap<RuleName, Rule<T>>, RuleError>>()?,
-        })
-    }
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-struct SerializableHistory {
-    steps: Vec<SerializableStep>,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct SerializableSimulation<T> {
-    history: SerializableHistory,
-    rules: HashSet<RuleName>,
-    possible_states: PossibleStates<T>,
-    cache: Cache,
-}
-
-#[derive(Clone, Default)]
-pub struct Simulation<T> {
-    history: Box<History<T>>,
-    rules: Box<HashMap<RuleName, Rule<T>>>,
-    possible_states: PossibleStates<T>,
-    cache: Cache,
-}
-
-impl<
-        T: Hash
-            + Clone
-            + PartialEq
-            + Debug
-            + Default
-            + Serialize
-            + Send
-            + Sync
-            + for<'a> Deserialize<'a>,
-    > Display for Simulation<T>
+impl<S, T> Debug for Simulation<S, T>
+where
+    S: Hash + Clone + Send + Sync + PartialEq + Debug,
+    T: Hash + Clone + Send + Sync + PartialEq + Debug,
 {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "Simulation:")?;
-        writeln!(f, "  Time: {}", self.time())?;
-        writeln!(f, "  Entropy: {}", self.entropy())?;
-        writeln!(f, "  Parameters:")?;
-        writeln!(f, "  Possible states:")?;
-        for (state_hash, state) in self.possible_states.iter() {
-            writeln!(f, "    {state_hash}: {state}")?;
-        }
-        writeln!(f, "  Reachable states:")?;
-        for (state_hash, probability) in self.reachable_states().iter() {
-            writeln!(f, "    {state_hash}: {probability}")?;
-        }
-        writeln!(f, "  Rules:")?;
-        for (rule_name, rule) in self.rules.iter() {
-            writeln!(f, "    {rule_name}: {rule}")?;
-        }
-        Ok(())
+        f.debug_struct("Simulation")
+            .field("state_transition_graph", &self.state_transition_graph)
+            .field("probabilities", &self.probability_distributions)
+            .field("known_states", &self.known_states)
+            .field("known_transitions", &self.known_transitions)
+            .finish()
     }
 }
 
-impl<T> Iterator for Simulation<T>
+impl<S, T> Simulation<S, T>
 where
-    T: Hash
-        + Clone
-        + PartialEq
-        + Debug
-        + Default
-        + Serialize
-        + Send
-        + Sync
-        + for<'a> Deserialize<'a>,
+    S: Hash + Clone + Send + Sync + PartialEq + Eq + Debug,
+    T: Hash + Clone + Send + Sync + PartialEq + Eq + Debug,
 {
-    type Item = Result<Simulation<T>, ErrorKind<T>>;
+    pub fn new(
+        initial_state: S,
+        state_transition_generator: StateTransitionGenerator<S, T>,
+    ) -> Self {
+        let initial_state_hash = hash(&initial_state);
 
-    fn next(&mut self) -> Option<Self::Item> {
-        let result = self.next_step();
-        match result {
-            Ok(_) => Some(Ok(self.clone())),
-            Err(error) => Some(Err(error)),
+        let mut state_transition_graph = Graph::new();
+        state_transition_graph.add_node(initial_state_hash);
+
+        let probabilities = HashMap::from([(0, HashMap::from([(initial_state_hash, 1.0)]))]);
+
+        let known_states = HashMap::from([(initial_state_hash, initial_state)]);
+
+        let known_transitions = HashMap::new();
+
+        Self {
+            state_transition_graph,
+            probability_distributions: probabilities,
+            known_states,
+            known_transitions,
+            state_transition_generator: CachedFunction::new(state_transition_generator),
         }
+    }
+
+    pub fn new_with_distribution(
+        probabilities: StateProbabilityDistribution<S>,
+        state_transition_generator: StateTransitionGenerator<S, T>,
+    ) -> Self {
+        let known_states = probabilities
+            .iter()
+            .map(|(state, _)| {
+                let state_hash = hash(state);
+                (state_hash, state.clone())
+            })
+            .collect::<HashMap<_, _>>();
+
+        let known_transitions = HashMap::new();
+
+        let hashed_probabilities = probabilities
+            .iter()
+            .map(|(state, probability)| {
+                let state_hash = hash(state);
+                (state_hash, *probability)
+            })
+            .collect::<HashMap<_, _>>();
+
+        let mut graph: StateTransitionGraph = Graph::new();
+        probabilities.iter().for_each(|(state, _)| {
+            let state_hash = hash(state);
+            graph.add_node(state_hash);
+        });
+
+        Self {
+            state_transition_graph: graph,
+            probability_distributions: HashMap::from([(0, hashed_probabilities)]),
+            known_states,
+            known_transitions,
+            state_transition_generator: CachedFunction::new(state_transition_generator),
+        }
+    }
+
+    fn state(&self, state_hash: StateHash) -> Option<&S> {
+        self.known_states.get(&state_hash)
+    }
+
+    fn transition(&self, transition_hash: TransitionHash) -> Option<&T> {
+        self.known_transitions.get(&transition_hash)
+    }
+
+    pub fn state_transition_graph(&self) -> Graph<S, (T, Probability)> {
+        let mut graph = Graph::new();
+        self.state_transition_graph
+            .node_indices()
+            .map(|node| {
+                let state_hash = *self.state_transition_graph.node_weight(node).unwrap();
+                self.state(state_hash).unwrap().clone()
+            })
+            .for_each(|state| {
+                graph.add_node(state);
+            });
+        let edges = self
+            .state_transition_graph
+            .edge_references()
+            .map(|edge| {
+                let source_hash = *self
+                    .state_transition_graph
+                    .node_weight(edge.source())
+                    .unwrap();
+                let target_hash = *self
+                    .state_transition_graph
+                    .node_weight(edge.target())
+                    .unwrap();
+                let (transition_hash, probability) = edge.weight();
+                let source = self.state(source_hash).unwrap().clone();
+                let target = self.state(target_hash).unwrap().clone();
+                let transition = self.transition(*transition_hash).unwrap().clone();
+                (source, target, transition, *probability)
+            })
+            .collect::<Vec<(S, S, T, Probability)>>();
+        for (source_state, target_state, transition, probability) in edges {
+            // check if nodes already exist
+            let source = graph
+                .node_indices()
+                .find(|node| graph.node_weight(*node).unwrap() == &source_state)
+                .unwrap_or_else(|| graph.add_node(source_state.clone()));
+            let target = graph
+                .node_indices()
+                .find(|node| graph.node_weight(*node).unwrap() == &target_state)
+                .unwrap_or_else(|| graph.add_node(target_state.clone()));
+            graph.update_edge(source, target, (transition, probability));
+        }
+        graph
+    }
+
+    pub fn probability_distributions(&self) -> HashMap<Time, StateProbabilityDistribution<S>> {
+        self.probability_distributions
+            .iter()
+            .map(|(time, state_probability_distribution)| {
+                let state_probability_distribution = state_probability_distribution
+                    .iter()
+                    .map(|(state_hash, probability)| {
+                        let state = self.state(*state_hash).unwrap().clone();
+                        (state, *probability)
+                    })
+                    .collect::<HashMap<_, _>>();
+
+                (*time, state_probability_distribution)
+            })
+            .collect::<HashMap<_, _>>()
+    }
+
+    pub fn state_probability(&self, state: S, time: Time) -> f64 {
+        self.probability_distributions
+            .get(&time)
+            .and_then(|state_probability_distribution| {
+                state_probability_distribution.get(&hash(&state))
+            })
+            .copied()
+            .unwrap_or(0.0)
+    }
+
+    pub fn initial_distribution(&self) -> StateProbabilityDistribution<S> {
+        self.probability_distribution(0)
+    }
+
+    pub fn probability_distribution(&self, time: Time) -> StateProbabilityDistribution<S> {
+        self.probability_distributions
+            .get(&time)
+            .map(|state_probability_distribution| {
+                state_probability_distribution
+                    .iter()
+                    .map(|(state_hash, probability)| {
+                        let state = self.state(*state_hash).unwrap().clone();
+                        (state, *probability)
+                    })
+                    .collect::<HashMap<_, _>>()
+            })
+            .expect("No probability distribution found for given time")
+    }
+
+    pub fn known_states(&self) -> Vec<S> {
+        self.known_states.values().cloned().collect()
+    }
+
+    pub fn known_transitions(&self) -> Vec<T> {
+        self.known_transitions.values().cloned().collect()
+    }
+
+    pub fn entropy(&self, time: Time) -> f64 {
+        let state_probability_distribution = self.probability_distribution(time);
+        let entropy = state_probability_distribution
+            .values()
+            .map(|probability| probability * probability.log2())
+            .sum::<f64>()
+            .abs();
+        entropy
+    }
+
+    pub fn time(&self) -> Time {
+        self.probability_distributions
+            .keys()
+            .max()
+            .copied()
+            .unwrap_or(0)
+    }
+
+    pub fn next_step(&mut self) -> StateProbabilityDistribution<S> {
+        let initial_time = self.time();
+        let state_probability_distribution: Vec<(S, Probability)> = self
+            .probability_distribution(initial_time)
+            .into_par_iter()
+            .collect();
+
+        let state_transition_probabilities = self.state_transition_generator.call_many_parallel(
+            state_probability_distribution
+                .par_iter()
+                .map(|(state, _)| state.clone()),
+        );
+
+        // Check if probabilities sum up to 1.0
+        state_transition_probabilities
+            .par_iter()
+            .for_each(|next_states| {
+                assert_eq!(
+                    (next_states
+                        .iter()
+                        .map(|(_, _, probability)| probability)
+                        .sum::<Probability>()
+                        * 10_i32.pow(10) as f64)
+                        .round()
+                        / 10_i32.pow(10) as f64,
+                    1.0,
+                    "Sum of probabilities of next states is not 1.0"
+                );
+            });
+
+        // Calculate new state probability distribution
+        let new_hashed_state_probability_distribution_mutex = Mutex::new(HashMap::new());
+        state_transition_probabilities
+            .par_iter()
+            .zip_eq(state_probability_distribution.par_iter())
+            .for_each(|(next_states, (_, current_state_probability))| {
+                next_states.iter().for_each(|(new_state, _, probability)| {
+                    new_hashed_state_probability_distribution_mutex
+                        .lock()
+                        .unwrap()
+                        .entry(hash(new_state))
+                        .and_modify(|state_probability| {
+                            *state_probability += current_state_probability * probability;
+                        })
+                        .or_insert(current_state_probability * probability);
+                });
+            });
+        // Add new state probability distribution to list of all state probability distributions
+        self.probability_distributions.insert(
+            initial_time + 1,
+            new_hashed_state_probability_distribution_mutex
+                .into_inner()
+                .unwrap(),
+        );
+
+        // Add new states and transitions to known states and transitions
+        state_transition_probabilities
+            .iter()
+            .for_each(|next_states| {
+                next_states.iter().for_each(|(new_state, transition, _)| {
+                    self.known_states.insert(hash(new_state), new_state.clone());
+                    self.known_transitions
+                        .insert(hash(transition), transition.clone());
+                });
+            });
+
+        // Add new state transitions to state transition graph
+        state_transition_probabilities
+            .iter()
+            .zip(state_probability_distribution.iter())
+            .for_each(|(next_states, (old_state, _))| {
+                next_states
+                    .iter()
+                    .for_each(|(new_state, transition, probability)| {
+                        let source = self
+                            .state_transition_graph
+                            .node_indices()
+                            .find(|node_index| {
+                                self.state_transition_graph
+                                    .node_weight(*node_index)
+                                    .unwrap()
+                                    == &hash(old_state)
+                            })
+                            .unwrap();
+                        let target = self
+                            .state_transition_graph
+                            .node_indices()
+                            .find(|node_index| {
+                                self.state_transition_graph
+                                    .node_weight(*node_index)
+                                    .unwrap()
+                                    == &hash(new_state)
+                            })
+                            .unwrap_or_else(|| {
+                                self.state_transition_graph.add_node(hash(new_state))
+                            });
+                        self.state_transition_graph.update_edge(
+                            source,
+                            target,
+                            (hash(transition), *probability),
+                        );
+                    });
+            });
+
+        // Return the new state probability distribution
+        self.probability_distribution(initial_time + 1)
+    }
+
+    pub fn full_traversal(&mut self, modify_cache_only: bool) {
+        if modify_cache_only {
+            let mut simulation_clone = self.clone();
+            let mut num_current_known_states = 0;
+            while num_current_known_states != simulation_clone.known_states.len() {
+                num_current_known_states = simulation_clone.known_states.len();
+                simulation_clone.next_step();
+                self.known_states = simulation_clone.known_states.clone();
+                self.known_transitions = simulation_clone.known_transitions.clone();
+                self.state_transition_graph = simulation_clone.state_transition_graph.clone();
+                self.state_transition_generator =
+                    simulation_clone.state_transition_generator.clone();
+            }
+        } else {
+            let mut num_current_known_states = 0;
+            while num_current_known_states != self.known_states.len() {
+                num_current_known_states = self.known_states.len();
+                self.next_step();
+            }
+        }
+    }
+
+    pub fn uniform_distribution_is_steady(&mut self) -> bool {
+        self.full_traversal(true);
+        let mut simulation_clone = self.clone();
+        let uniform_probability = 1.0 / self.known_states.len() as Probability;
+        let uniform_state_probability_distribution = self
+            .known_states
+            .iter()
+            .map(|(state_hash, _)| (*state_hash, uniform_probability))
+            .collect::<HashMap<_, _>>();
+        let next_time = simulation_clone.time() + 1;
+        simulation_clone
+            .probability_distributions
+            .insert(next_time, uniform_state_probability_distribution);
+        let uniform_entropy = simulation_clone.entropy(next_time);
+        simulation_clone.next_step();
+        let uniform_entropy_after_step = simulation_clone.entropy(next_time + 1);
+        uniform_entropy_after_step == uniform_entropy
     }
 }
 
-impl<
-        T: Hash
-            + Clone
-            + PartialEq
-            + Debug
-            + Default
-            + Serialize
-            + Send
-            + Sync
-            + for<'a> Deserialize<'a>,
-    > Simulation<T>
-{
-    pub fn new(initial_state: State<T>, rules: HashMap<RuleName, Rule<T>>) -> Simulation<T> {
-        let initial_state_hash = StateHash::new(&initial_state);
-        Simulation::new_with_reachable_states(
-            PossibleStates::new(HashMap::from([(initial_state_hash, initial_state)])),
-            ReachableStates::from(HashMap::from([(initial_state_hash, Probability::from(1.))])),
-            rules,
-        )
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn random_walk() {
+        let initial_state = 0;
+        let state_transition_generator =
+            Arc::new(|state: i32| vec![(state + 1, "next", 0.5), (state - 1, "previous", 0.5)]);
+
+        let mut simulation = Simulation::new(initial_state, state_transition_generator);
+        dbg!(&simulation);
+        assert_eq!(simulation.known_states().len(), 1);
+        assert_eq!(simulation.known_transitions().len(), 0);
+        assert_eq!(simulation.probability_distributions().len(), 1);
+        assert_eq!(simulation.state_transition_graph().node_count(), 1);
+        assert_eq!(simulation.state_transition_graph().edge_count(), 0);
+        assert_eq!(simulation.entropy(0), 0.0);
+
+        simulation.next_step();
+        dbg!(&simulation);
+        assert_eq!(simulation.known_states().len(), 3);
+        assert_eq!(simulation.known_transitions().len(), 2);
+        assert_eq!(simulation.probability_distributions().len(), 2);
+        assert_eq!(simulation.state_transition_graph().node_count(), 3);
+        assert_eq!(simulation.state_transition_graph().edge_count(), 2);
+        assert_eq!(simulation.entropy(1), 1.0);
+
+        let graph = simulation.state_transition_graph();
+        dbg!(&graph);
+        assert_eq!(graph.node_count(), 3);
+        assert_eq!(graph.edge_count(), 2);
+
+        assert_eq!(simulation.state_probability(0, 1), 0.);
+        assert_eq!(simulation.state_probability(1, 1), 0.5);
+        assert_eq!(simulation.state_probability(-1, 1), 0.5);
+
+        assert_eq!(simulation.initial_distribution(), HashMap::from([(0, 1.0)]));
+
+        assert_eq!(simulation.time(), 1);
     }
 
-    pub fn new_with_reachable_states(
-        possible_states: PossibleStates<T>,
-        reachable_states: ReachableStates,
-        rules: HashMap<RuleName, Rule<T>>,
-    ) -> Simulation<T> {
-        Simulation {
-            possible_states,
-            rules: Box::new(rules),
-            history: Box::new(History::new(reachable_states)),
-            cache: Cache::new(),
-        }
+    #[test]
+    fn random_walk_with_initial_distribution() {
+        let initial_distribution = HashMap::from([(0, 0.5), (1, 0.5)]);
+        let state_transition_generator =
+            Arc::new(|state: i32| vec![(state + 1, "next", 0.5), (state - 1, "previous", 0.5)]);
+        let mut simulation =
+            Simulation::new_with_distribution(initial_distribution, state_transition_generator);
+        assert_eq!(simulation.known_states().len(), 2);
+        assert_eq!(simulation.known_transitions().len(), 0);
+        assert_eq!(simulation.probability_distributions().len(), 1);
+        assert_eq!(simulation.state_transition_graph().node_count(), 2);
+        assert_eq!(simulation.state_transition_graph().edge_count(), 0);
+        assert_eq!(simulation.entropy(0), 1.0);
+        dbg!(&simulation);
+
+        simulation.next_step();
+
+        assert_eq!(simulation.known_states().len(), 4);
+        assert_eq!(simulation.known_transitions().len(), 2);
+        assert_eq!(simulation.probability_distributions().len(), 2);
+        assert_eq!(simulation.state_transition_graph().node_count(), 4);
+        assert_eq!(simulation.state_transition_graph().edge_count(), 4);
+        assert_eq!(simulation.entropy(1), 2.0);
+        assert_eq!(
+            simulation.probability_distributions(),
+            HashMap::from([
+                (0, HashMap::from([(0, 0.5), (1, 0.5)]),),
+                (
+                    1,
+                    HashMap::from([(-1, 0.25), (0, 0.25), (1, 0.25), (2, 0.25)]),
+                ),
+            ])
+        );
+
+        dbg!(&simulation);
     }
 
-    pub fn history(&self) -> &History<T> {
-        &self.history
-    }
-
-    pub fn to_serializable(&self) -> SerializableSimulation<T> {
-        let history = SerializableHistory {
-            steps: self
-                .history
-                .steps()
-                .iter()
-                .map(|step| SerializableStep {
-                    reachable_states: step.reachable_states().clone(),
-                    applied_rules: step.applied_rules().keys().cloned().collect(),
-                })
-                .collect(),
-        };
-        SerializableSimulation {
-            history,
-            rules: self.rules.keys().cloned().collect(),
-            possible_states: self.possible_states.clone(),
-            cache: self.cache.clone(),
-        }
-    }
-
-    pub fn from_serializable(
-        serializable_simulation: SerializableSimulation<T>,
-        rules: HashMap<RuleName, Rule<T>>,
-    ) -> Result<Simulation<T>, ErrorKind<T>> {
-        let steps: Vec<Step<T>> = serializable_simulation
-            .history
-            .steps
-            .iter()
-            .map(|step| step.to_step::<T>(&rules))
-            .collect::<Result<Vec<Step<T>>, RuleError>>()?;
-        let history = History { steps };
-        Ok(Simulation {
-            history: Box::new(history),
-            rules: Box::new(
-                serializable_simulation
-                    .rules
-                    .iter()
-                    .map(|rule_name| {
-                        let rule = rules.get(rule_name);
-                        if let Some(rule) = rule {
-                            Ok((rule_name.clone(), rule.clone()))
+    #[test]
+    fn full_traversal() {
+        let initial_state = 0;
+        const NUM_STATES: i32 = 5;
+        let state_transition_generator = Arc::new(|state: i32| -> OutgoingTransitions<i32, &str> {
+            vec![
+                (
+                    {
+                        if state + 1 == NUM_STATES {
+                            0
                         } else {
-                            Err(RuleError::RuleNotFound {
-                                rule_name: rule_name.clone(),
-                                context: get_backtrace(),
-                            })
+                            state + 1
                         }
-                    })
-                    .collect::<Result<HashMap<RuleName, Rule<T>>, RuleError>>()?,
-            ),
-            possible_states: serializable_simulation.possible_states,
-            cache: serializable_simulation.cache,
-        })
+                    },
+                    "forward",
+                    0.5,
+                ),
+                (
+                    {
+                        if state - 1 == -1 {
+                            NUM_STATES - 1
+                        } else {
+                            state - 1
+                        }
+                    },
+                    "backward",
+                    0.5,
+                ),
+            ]
+        });
+        let mut simulation = Simulation::new(initial_state, state_transition_generator);
+        simulation.full_traversal(false);
+        dbg!(&simulation);
+        let graph = simulation.state_transition_graph();
+        let dot = petgraph::dot::Dot::with_config(&graph, &[]);
+        println!("{dot:#?}");
+        assert_eq!(simulation.known_states().len(), NUM_STATES as usize);
+        assert_eq!(simulation.known_transitions().len(), 2);
+        assert_eq!(
+            simulation.probability_distribution(simulation.time()).len(),
+            4
+        );
+        assert_eq!(
+            simulation.state_transition_graph().node_count(),
+            NUM_STATES as usize
+        );
+        assert_eq!(
+            simulation.state_transition_graph().edge_count(),
+            2 * NUM_STATES as usize
+        );
     }
 
-    pub fn clone_without_history(&self) -> Simulation<T> {
-        Simulation {
-            history: Box::new(History::new(self.initial_distribution().clone())),
-            rules: self.rules.clone(),
-            possible_states: self.possible_states.clone(),
-            cache: self.cache.clone(),
+    #[test]
+    fn uniform_distribution_is_steady() {
+        {
+            let initial_state = 0;
+            const NUM_STATES: i32 = 5;
+            let state_transition_generator =
+                Arc::new(|state: i32| -> OutgoingTransitions<i32, &str> {
+                    vec![
+                        (
+                            {
+                                if state + 1 == NUM_STATES {
+                                    0
+                                } else {
+                                    state + 1
+                                }
+                            },
+                            "forward",
+                            0.5,
+                        ),
+                        (
+                            {
+                                if state - 1 == -1 {
+                                    NUM_STATES - 1
+                                } else {
+                                    state - 1
+                                }
+                            },
+                            "backward",
+                            0.5,
+                        ),
+                    ]
+                });
+            let mut simulation = Simulation::new(initial_state, state_transition_generator);
+            assert!(simulation.uniform_distribution_is_steady());
         }
-    }
-
-    pub fn initial_distribution(&self) -> &ReachableStates {
-        self.history.steps().first().unwrap().reachable_states()
-    }
-
-    pub fn possible_states(&self) -> &PossibleStates<T> {
-        &self.possible_states
-    }
-
-    pub fn reachable_states(&self) -> &ReachableStates {
-        self.history().steps().last().unwrap().reachable_states()
-    }
-
-    pub fn next_step_with_distribution(&mut self, reachable_states: ReachableStates) {
-        self.history
-            .append(Step::new(reachable_states, HashMap::new()));
-    }
-
-    pub fn rules(&self) -> &HashMap<RuleName, Rule<T>> {
-        &self.rules
-    }
-
-    pub fn time(&self) -> usize {
-        self.history().steps().len() - 1
-    }
-
-    pub fn entropy(&self) -> Entropy {
-        self.reachable_states().entropy()
-    }
-
-    pub fn next_step(&mut self) -> Result<(), ErrorKind<T>> {
-        let rules = self.rules.clone();
-        let next_reachable_states = self.next_reachable_states(&rules)?;
-        self.history
-            .append(Step::new(next_reachable_states, *rules));
-        Ok(())
-    }
-
-    pub fn run(&mut self, steps: usize) -> Result<(), ErrorKind<T>> {
-        for _ in 0..steps {
-            self.next_step()?;
+        {
+            let initial_state = 0;
+            const NUM_STATES: i32 = 5;
+            let state_transition_generator =
+                Arc::new(|state: i32| -> OutgoingTransitions<i32, &str> {
+                    vec![
+                        (
+                            {
+                                if state + 1 == NUM_STATES {
+                                    1
+                                } else {
+                                    state + 1
+                                }
+                            },
+                            "forward",
+                            0.5,
+                        ),
+                        (0, "stay", 0.5),
+                    ]
+                });
+            let mut simulation = Simulation::new(initial_state, state_transition_generator);
+            assert!(!simulation.uniform_distribution_is_steady());
         }
-        Ok(())
-    }
-
-    /// Performs a full traversal of the possible states of the system.
-    ///
-    /// This method will continue to call `next_step()` until all possible states have been visited.
-    /// If an `iteration_limit` is provided, the traversal will stop if the time spent exceeds the limit.
-    /// If modify_state is set to false, only the cache and the possible states will be updated,
-    /// but the simulation will otherwise remain at its current state.
-    ///
-    /// # Errors
-    ///
-    /// - `ErrorKind::IterationLimitReached` - If the traversal took longer than the provided iteration limit.
-    ///   Note that any progress will be  applied to the simulation.
-    /// ```
-    pub fn full_traversal(
-        &mut self,
-        iteration_limit: Option<usize>,
-        modify_state: bool,
-    ) -> Result<(), ErrorKind<T>> {
-        if modify_state {
-            let mut num_current_possible_states = 0;
-            while num_current_possible_states != self.possible_states().len() {
-                if let Some(iteration_limit) = iteration_limit {
-                    if self.time() >= iteration_limit {
-                        return Err(ErrorKind::IterationLimitReached {
-                            time: self.time(),
-                            context: get_backtrace(),
-                        });
-                    }
-                }
-                num_current_possible_states = self.possible_states().len();
-                self.next_step()?;
-            }
-            Ok(())
-        } else {
-            let mut simulation_clone = self.clone();
-            simulation_clone.full_traversal(iteration_limit, true)?;
-            self.possible_states
-                .merge(simulation_clone.possible_states())?;
-            self.cache.merge(&simulation_clone.cache)?;
-            Ok(())
-        }
-    }
-
-    pub fn apply_intervention(
-        &mut self,
-        rules: &HashMap<RuleName, Rule<T>>,
-    ) -> Result<(), ErrorKind<T>> {
-        for rule_name in rules.keys() {
-            if self.rules.contains_key(rule_name) {
-                return Err(ErrorKind::from(CacheError::RuleAlreadyExists {
-                    rule_name: rule_name.clone(),
-                    context: get_backtrace(),
-                }));
-            }
-        }
-        let next_reachable_states = self.next_reachable_states(rules)?;
-        self.history
-            .append(Step::new(next_reachable_states, rules.clone()));
-        Ok(())
-    }
-
-    fn next_reachable_states(
-        &mut self,
-        rules: &HashMap<RuleName, Rule<T>>,
-    ) -> Result<ReachableStates, ErrorKind<T>> {
-        self.reachable_states().clone().apply_rules(
-            &mut self.possible_states,
-            &mut self.cache,
-            rules,
-        )
-    }
-
-    pub fn graph(
-        &mut self,
-        iteration_limit: Option<usize>,
-    ) -> Result<Graph<StateHash, RuleName>, ErrorKind<T>> {
-        self.full_traversal(iteration_limit, false)?;
-        self.cache.graph(self.possible_states.clone())
-    }
-
-    pub fn uniform_distribution_is_steady(
-        &mut self,
-        iteration_limit: Option<usize>,
-    ) -> Result<bool, ErrorKind<T>> {
-        self.full_traversal(iteration_limit, false)?;
-        let simulation = self.clone();
-        let uniform_probability = Probability::from(1. / simulation.possible_states.len() as f64);
-        let uniform_distribution: ReachableStates = ReachableStates::from(HashMap::from_iter(
-            simulation.possible_states.iter().map(|(state_hash, _)| {
-                let prob: (StateHash, Probability) = (*state_hash, uniform_probability);
-                prob
-            }),
-        ));
-        let mut uniform_simulation = simulation;
-        uniform_simulation.next_step_with_distribution(uniform_distribution);
-        let uniform_entropy = uniform_simulation.entropy();
-        uniform_simulation.next_step()?;
-        let uniform_entropy_after_step = uniform_simulation.entropy();
-        Ok(uniform_entropy == uniform_entropy_after_step)
     }
 }
